Fix repeat value type in Quran audio controller

+++ flutter_app/lib/features/quran/quran_audio_controller.dart
@@ -93,7 +93,7 @@ class QuranAudioController extends ChangeNotifier {
   }
 
   void setRepeat(int value) {
-    _repeat = value.clamp(1, 51);
+    _repeat = value.clamp(1, 51).toInt();
     notifyListeners();
   }
